cql3/external_search: extract `external_index_select_statement` base class
Factor out the base-table fetch machinery shared between ANN and FTS
SELECT statements into a new `external_index_select_statement` base class.

`vector_indexed_table_select_statement` and
`fulltext_indexed_table_select_statement` now both inherit from this base.
