+++ configure.py
@@ -1065,6 +1065,7 @@ scylla_core = (['message/messaging_service.cc',
                 'cql3/statements/prune_materialized_view_statement.cc',
                 'cql3/statements/batch_statement.cc',
                 'cql3/statements/select_statement.cc',
+                'cql3/statements/external_search/external_index_select_statement.cc',
                 'cql3/statements/external_search/vector_indexed_table_select_statement.cc',
                 'cql3/statements/external_search/fulltext_indexed_table_select_statement.cc',
                 'cql3/statements/external_search/filter.cc',
