+++ cql3/statements/external_search/vector_indexed_table_select_statement.cc
@@ -19,18 +19,12 @@
 #include "replica/database.hh"
 #include "exceptions/exceptions.hh"
 #include "index/vector_index.hh"
-#include "query/query_result_merger.hh"
-#include "service/storage_proxy.hh"
 #include "types/vector.hh"
-#include "utils/result_loop.hh"
 
 #include <seastar/core/future.hh>
 #include <seastar/core/on_internal_error.hh>
 #include <seastar/coroutine/exception.hh>
 
-
-template<typename T = void>
-using coordinator_result = cql3::statements::select_statement::coordinator_result<T>;
 
 namespace cql3 {
 
@@ -52,37 +46,6 @@ auto measure_index_latency(const schema& schema, const secondary_index::index& i
     }
 
     co_return result;
-}
-
-template<typename C>
-struct result_to_error_message_wrapper {
-    C c;
-
-    template<typename T>
-    auto operator()(coordinator_result<T>&& arg) {
-        if constexpr (std::is_void_v<T>) {
-            if (arg) {
-                return futurize_invoke(c);
-            } else {
-                return make_ready_future<typename futurize_t<std::invoke_result_t<C>>::value_type>(
-                    ::make_shared<cql_transport::messages::result_message::exception>(std::move(arg).assume_error())
-                );
-            }
-        } else {
-            if (arg) {
-                return futurize_invoke(c, std::move(arg).value());
-            } else {
-                return make_ready_future<typename futurize_t<std::invoke_result_t<C, T>>::value_type>(
-                    ::make_shared<cql_transport::messages::result_message::exception>(std::move(arg).assume_error())
-                );
-            }
-        }
-    }
-};
-
-template<typename C>
-auto wrap_result_to_error_message(C&& c) {
-    return result_to_error_message_wrapper<C>{std::move(c)};
 }
 
 std::vector<float> get_ann_ordering_vector(const select_statement::prepared_ann_ordering_type& prepared_ann_ordering, const query_options& options) {
@@ -219,9 +182,8 @@ vector_indexed_table_select_statement::vector_indexed_table_select_statement(sch
         prepared_ann_ordering_type prepared_ann_ordering, std::optional<expr::expression> limit,
         std::optional<expr::expression> per_partition_limit, cql_stats& stats, const secondary_index::index& index,
         external_search::prepared_filter prepared_filter, std::unique_ptr<attributes> attrs)
-    : select_statement{schema, bound_terms, parameters, selection, restrictions, group_by_cell_indices, is_reversed, ordering_comparator, limit,
-              per_partition_limit, stats, std::move(attrs)}
-    , _index{index}
+    : external_index_select_statement{schema, bound_terms, parameters, selection, restrictions, group_by_cell_indices,
+              is_reversed, ordering_comparator, limit, per_partition_limit, stats, index, std::move(attrs)}
     , _prepared_ann_ordering(std::move(prepared_ann_ordering))
     , _prepared_filter(std::move(prepared_filter)) {
 
@@ -279,83 +241,6 @@ future<shared_ptr<cql_transport::messages::result_message>> vector_indexed_table
         result->add_warning("Paging is not supported for Vector Search queries. The entire result set has been returned.");
     }
     co_return result;
-}
-
-void vector_indexed_table_select_statement::update_stats() const {
-    ++_stats.secondary_index_reads;
-    ++_stats.query_cnt(source_selector::USER, _ks_sel, cond_selector::NO_CONDITIONS, statement_type::SELECT);
-}
-
-lw_shared_ptr<query::read_command> vector_indexed_table_select_statement::prepare_command_for_base_query(
-        query_processor& qp, service::query_state& state, const query_options& options, uint64_t fetch_limit) const {
-    auto slice = make_partition_slice(options);
-    return ::make_lw_shared<query::read_command>(_schema->id(), _schema->version(), std::move(slice), qp.proxy().get_max_result_size(slice),
-            query::tombstone_limit(qp.proxy().get_tombstone_limit()),
-            query::row_limit(get_inner_loop_limit(fetch_limit, _selection->is_aggregate())), query::partition_limit(query::max_partitions),
-            _query_start_time_point, tracing::make_trace_info(state.get_trace_state()), query_id::create_null_id(), query::is_first_page::no,
-            options.get_timestamp(state));
-}
-
-future<::shared_ptr<cql_transport::messages::result_message>> vector_indexed_table_select_statement::query_base_table(query_processor& qp,
-        service::query_state& state, const query_options& options, const std::vector<vector_search::primary_key>& pkeys,
-        lowres_clock::time_point timeout) const {
-    auto command = prepare_command_for_base_query(qp, state, options, pkeys.size());
-
-    auto result = co_await query_base_table(qp, state, options, command, timeout, pkeys);
-
-    command->set_row_limit(get_limit(options, _limit));
-
-    co_return co_await wrap_result_to_error_message([this, command = std::move(command), &options](auto query_result) {
-        return process_results(std::move(query_result), command, options, _query_start_time_point);
-    })(std::move(result));
-}
-
-future<coordinator_result<foreign_ptr<lw_shared_ptr<query::result>>>> vector_indexed_table_select_statement::query_base_table(query_processor& qp,
-        service::query_state& state, const query_options& options, lw_shared_ptr<query::read_command> command, lowres_clock::time_point timeout,
-        const std::vector<vector_search::primary_key>& pkeys) const {
-
-    // For tables without clustering columns, we can optimize by querying
-    // partition ranges instead of individual primary keys, since the
-    // partition key alone uniquely identifies each row.
-    if (_schema->clustering_key_size() == 0) {
-        auto to_partition_ranges = [](const std::vector<vector_search::primary_key>& pkeys) -> std::vector<dht::partition_range> {
-            std::vector<dht::partition_range> partition_ranges;
-            std::ranges::transform(pkeys, std::back_inserter(partition_ranges), [](const auto& pkey) {
-                return dht::partition_range::make_singular(pkey.partition);
-            });
-
-            return partition_ranges;
-        };
-        co_return co_await query_base_table(qp, state, options, std::move(command), timeout, to_partition_ranges(pkeys));
-    }
-    co_return co_await utils::result_map_reduce(
-            pkeys.begin(), pkeys.end(),
-            [&](this auto, auto& key) -> future<coordinator_result<foreign_ptr<lw_shared_ptr<query::result>>>> {
-                auto cmd = ::make_lw_shared<query::read_command>(*command);
-                cmd->slice._row_ranges = query::clustering_row_ranges{query::clustering_range::make_singular(key.clustering)};
-                coordinator_result<service::storage_proxy::coordinator_query_result> rqr =
-                        co_await qp.proxy().query_result(_schema, cmd, {dht::partition_range::make_singular(key.partition)}, options.get_consistency(),
-                                {timeout, state.get_permit(), state.get_client_state(), state.get_trace_state()});
-                if (!rqr) {
-                    co_return std::move(rqr).as_failure();
-                }
-                co_return std::move(rqr.value().query_result);
-            },
-            query::result_merger{command->get_row_limit(), query::max_partitions});
-}
-
-future<coordinator_result<foreign_ptr<lw_shared_ptr<query::result>>>> vector_indexed_table_select_statement::query_base_table(query_processor& qp,
-        service::query_state& state, const query_options& options, lw_shared_ptr<query::read_command> command, lowres_clock::time_point timeout,
-        std::vector<dht::partition_range> partition_ranges) const {
-
-    coordinator_result<service::storage_proxy::coordinator_query_result> rqr = co_await qp.proxy()
-            .query_result(_query_schema, command, std::move(partition_ranges), options.get_consistency(),
-                    {timeout, state.get_permit(), state.get_client_state(), state.get_trace_state(), {}, {}, options.get_specific_options().node_local_only},
-                    std::nullopt);
-    if (!rqr) {
-        co_return std::move(rqr).as_failure();
-    }
-    co_return std::move(rqr.value().query_result);
 }
 
 } // namespace statements
